Swipe to show swipe menu not working in Wallets screen opened from Wallet tab #2916
