+++ AlphaWallet/AppDelegate.swift
@@ -1,6 +1,5 @@
 // Copyright SIX DAY LLC. All rights reserved.
-import UIKit
-import RealmSwift
+import UIKit 
 import AWSSNS
 import AWSCore
 import UserNotifications
