+++ AlphaWallet/Settings/Coordinators/SettingsCoordinator.swift
@@ -207,9 +207,7 @@ extension SettingsCoordinator: AccountsCoordinatorDelegate {
 
 	func didDeleteAccount(account: Wallet, in coordinator: AccountsCoordinator) {
         delegate?.delete(account: account, in: self)
-        for each in sessions.values {
-			TransactionsTracker(sessionID: each.sessionID).fetchingState = .initial
-		}
+        TransactionsTracker.functional.resetFetchingState(account: account, config: config)
 		delegate?.didUpdateAccounts(in: self)
 		guard !coordinator.accountsViewController.hasWallets else { return }
         coordinator.navigationController.popViewController(animated: true)
