+++ AlphaWallet/Tokens/Types/TokenObject.swift
@@ -305,3 +305,16 @@ func compositeTokenName(forContract contract: AlphaWallet.Address, fromContractN
     }
     return compositeName
 }
+
+
+extension Wallet {
+    class functional {}
+}
+
+extension Wallet.functional {
+    static func realm(forAccount account: Wallet) -> Realm {
+        let migration = MigrationInitializer(account: account)
+        migration.perform()
+        return try! Realm(configuration: migration.config)
+    }
+}
